Remove old route
