+++ pgdog/src/frontend/router/parser/mod.rs
@@ -3,8 +3,10 @@
 pub mod comment;
 pub mod copy;
 pub mod error;
+pub mod order_by;
 pub mod query;
 pub mod route;
 
 pub use error::Error;
+pub use order_by::OrderBy;
 pub use route::Route;
